Add SmallRuby
I'm adding SmallRuby, an university research project here at Faculty of Information Technologies at Czech Technical University in Prague, Czech Republic. It's been dead for a while, but we have hopes to resurrect it and play with it more (we're working on the parser right now). But until then, I would list it as discontinued ruby. Thanks!